requirements.txt
fastapi
sqlalchemy
uvicorn
pymysql
cryptography
scipy
pandas
numpy
psycopg2
python-ulid
"passlib[bcrypt]"
python-dotenv
"python-jose[cryptography]"
python-multipart